logic-api/api/requirements.txt
Flask==2.0.1
textblob==0.15.0
py-healthcheck==1.10.1